+++ src/main/java/dev/boze/client/mixin/PlayerInventoryMixin.java
@@ -9,14 +9,14 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin({PlayerInventory.class})
 public class PlayerInventoryMixin {
-   @Inject(
-      method = {"getBlockBreakingSpeed"},
-      at = {@At("RETURN")},
-      cancellable = true
-   )
-   private void onGetBlockBreakingSpeed(CallbackInfoReturnable<Float> var1) {
-      if (MiningTweaks.INSTANCE.isEnabled()) {
-         var1.setReturnValue((Float)var1.getReturnValue() * MiningTweaks.INSTANCE.speed.getValue());
-      }
-   }
+    @Inject(
+            method = {"getBlockBreakingSpeed"},
+            at = {@At("RETURN")},
+            cancellable = true
+    )
+    private void onGetBlockBreakingSpeed(CallbackInfoReturnable<Float> var1) {
+        if (MiningTweaks.INSTANCE.isEnabled()) {
+            var1.setReturnValue(var1.getReturnValue() * MiningTweaks.INSTANCE.speed.getValue());
+        }
+    }
 }
