+++ src/main/java/dev/boze/client/mixin/Vec3dMixin.java
@@ -11,38 +11,38 @@ import org.spongepowered.asm.mixin.Shadow;
 
 @Mixin({Vec3d.class})
 public class Vec3dMixin implements IVec3d {
-   @Shadow
-   @Final
-   @Mutable
-   public double x;
-   @Shadow
-   @Final
-   @Mutable
-   public double y;
-   @Shadow
-   @Final
-   @Mutable
-   public double z;
+    @Shadow
+    @Final
+    @Mutable
+    public double x;
+    @Shadow
+    @Final
+    @Mutable
+    public double y;
+    @Shadow
+    @Final
+    @Mutable
+    public double z;
 
-   @Override
-   public void boze$set(double x, double y, double z) {
-      this.x = x;
-      this.y = y;
-      this.z = z;
-   }
+    @Override
+    public void boze$set(double x, double y, double z) {
+        this.x = x;
+        this.y = y;
+        this.z = z;
+    }
 
-   @Override
-   public void boze$set(Vec3i vec) {
-      this.boze$set((double)vec.getX(), (double)vec.getY(), (double)vec.getZ());
-   }
+    @Override
+    public void boze$set(Vec3i vec) {
+        this.boze$set(vec.getX(), vec.getY(), vec.getZ());
+    }
 
-   @Override
-   public void boze$set(Vector3d vec) {
-      this.boze$set(vec.x, vec.y, vec.z);
-   }
+    @Override
+    public void boze$set(Vector3d vec) {
+        this.boze$set(vec.x, vec.y, vec.z);
+    }
 
-   @Override
-   public void boze$setY(double y) {
-      this.y = y;
-   }
+    @Override
+    public void boze$setY(double y) {
+        this.y = y;
+    }
 }
